TPG - Rethinking Rx
- Instead of allocating thousand of small counters per tag separately when TPG is enabled, allocate them together as a big chunk. Do this using calloc. (This might fragment)

- This improves performance since zeroing the memory is distributed by the COW (Copy on Write) mechanism.

- Allocate the counters in another thread so that the RX can have high availability to support multiple users.

- Convert the state machine syncing mechanism to use messaging instead of shared memory now that the Rx is available to receive messages.

+++ src/stx/stl/trex_stl_rpc_cmds.cpp
@@ -1793,7 +1793,8 @@ TrexRpcCmdGetTaggedPktGroupState::_run(const Json::Value &params, Json::Value &r
 
     TPGState state = stl->update_tpg_state(username);
 
-    result["result"] = state;
+    // TPGState is an enum class, hence we need to cast it back to an integer.
+    result["result"] = static_cast<int>(state);
 
     return (TREX_RPC_CMD_OK);
 }
